moved logic from main script to imported module + removed compiled js files
