+++ package.json
@@ -2,21 +2,21 @@
   "name": "pegasus-feeds",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/jjordan-quantum/pegasus-feeds.git"
+    "url": "git+https://github.com/umbrella-network/pegasus-feeds.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/jjordan-quantum/pegasus-feeds/issues"
+    "url": "https://github.com/umbrella-network/pegasus-feeds/issues"
   },
-  "homepage": "https://github.com/jjordan-quantum/pegasus-feeds#readme",
+  "homepage": "https://github.com/umbrella-network/pegasus-feeds#readme",
   "devDependencies": {
     "@types/node": "^17.0.8"
   },
