--- src/main.js
@@ -1,92 +1,0 @@
-"use strict";
-var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
-    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
-exports.__esModule = true;
-var add_feeds_1 = require("./add-feeds");
-/*
- *  addFeeds
- *
- *  Will add symbols found in files within the /add/ directory as new keys in dev/bsc/feeds.yaml.
- *  Files must contain 1 symbol per line.
- *  Duplicates are ignored.
- *  Each file with valid symbols will create a new section in the yaml, using the filename as heading (in a comment).
- *
- */
-function addFeeds() {
-    return __awaiter(this, void 0, void 0, function () {
-        var DISCREPANCY, PRECISION, FETCHER, YAML_FILE_PATH, ADD_DIR_PATH, feedYamlData;
-        return __generator(this, function (_a) {
-            DISCREPANCY = 1.0;
-            PRECISION = 2;
-            FETCHER = "PolygonIOPrice";
-            YAML_FILE_PATH = "../dev/bsc/feeds.yaml";
-            ADD_DIR_PATH = "../add/";
-            feedYamlData = (0, add_feeds_1.readFeedsYaml)(YAML_FILE_PATH);
-            // step 2 - add symbols from each file, if required
-            (0, add_feeds_1.getAllFiles)(ADD_DIR_PATH).forEach(function (filename) {
-                var feedBuffer = '\n' + '# ' + filename + '\n';
-                (0, add_feeds_1.getAllSymbols)(ADD_DIR_PATH + filename).forEach(function (symbol) {
-                    var feedSymbol = 'EQ:' + symbol;
-                    if (!feedYamlData.hasOwnProperty(feedSymbol)) {
-                        feedBuffer += '\n' + (0, add_feeds_1.getEquityYamlDump)(symbol, DISCREPANCY, PRECISION, FETCHER);
-                    }
-                });
-                // count \n's, if greater than 2 -> dump
-                if ((feedBuffer.split("\n").length - 1) > 2) {
-                    (0, add_feeds_1.appendFeedsYaml)(YAML_FILE_PATH, feedBuffer);
-                }
-            });
-            return [2 /*return*/];
-        });
-    });
-}
-(function () {
-    return __awaiter(this, void 0, void 0, function () {
-        return __generator(this, function (_a) {
-            switch (_a.label) {
-                case 0: return [4 /*yield*/, addFeeds()];
-                case 1:
-                    _a.sent();
-                    return [2 /*return*/];
-            }
-        });
-    });
-})()
-    .then(function () { process.exit(0); })["catch"](function (error) {
-    console.error(error);
-    process.exit(1);
-});
